requirements.txt
ipykernel
openai
langchain
pyyaml
streamlit
